created a separate attendance mobile component
created a separate attendance mobile component
