+++ src/app/feature_modules/dashboard/dashboard.module.ts
@@ -3,6 +3,7 @@ import { CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA, NgModule } from "@angular/cor
 import { SharedModule } from "src/app/shared/shared.Module";
 import { DashboardComponent } from './dashboard.component';
 import { AttendanceLogsDialogComponent } from './attendance-logs-dialog.component';
+import { AttendanceOverviewComponent } from './attendance-overview/attendance-overview.component';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatDatepickerModule } from '@angular/material/datepicker';
 import { MatButtonModule } from '@angular/material/button';
@@ -14,7 +15,7 @@ import { CommonModule } from '@angular/common';
 import { DashboardRoutingModule } from './dashboard.routing.module';
 
 @NgModule({
-  declarations: [UserDashboardComponent, DashboardComponent, AttendanceLogsDialogComponent],
+  declarations: [UserDashboardComponent, DashboardComponent, AttendanceLogsDialogComponent, AttendanceOverviewComponent],
   exports: [UserDashboardComponent, DashboardComponent],
   imports: [
     DashboardRoutingModule,
